Correctly determine centos kernel package
The CentOS kernel package (kernel-0:<number>.el?) is
a dummy package that's nog included in the RHA lists.
To prevent false positives, these packages need to be
filtered from the list.

Example:

For an RHA, kernel needs to be version 3.10.0-862.11.6.el7
rpm with name kernel-3.10.0-862.el7 would be to old and
trigger a warning about this RHA, even when the correct
rpm kernel-3.10.0-862.11.6.el7 is installed.

To prevent these issues, we filter all kernel packages
without a minor in the release. (ie release is ####.el?)

These packages are selected by splitting the release and
comparing the 1st 2 characters of the 2nd field with el.
To ensure we're not comparing a shorter string, the part
we compare must at least consists of 2 characters. (else
it wouldn't fit 'el' and would pass anyway)